create test for student assignment api

+++ src/routes/studentAssignment-router.ts
@@ -19,6 +19,12 @@ StudentAssignmentRouter
       },
     })
 
+    if (!studentAssignment) {
+      return c.json({
+        message: 'Student assignment not found',
+      }, 404)
+    }
+
     return c.json({ student_assignment: studentAssignment })
   })
 
@@ -39,7 +45,10 @@ StudentAssignmentRouter
       },
     })
 
-    return c.json({ student_assignment: studentAssignment })
+    return c.json({
+      message: 'Created',
+      student_assignment: studentAssignment,
+    })
   })
 
   // PUT /api/student-assignments/:id
@@ -51,7 +60,19 @@ StudentAssignmentRouter
       deadline_date,
       subject_id,
     } = await c.req.json()
-    const studentAssignment = await prisma.student_assignments.update({
+    let studentAssignment = await prisma.student_assignments.findFirst({
+      where: {
+        id: Number.parseInt(id),
+      },
+    })
+
+    if (!studentAssignment) {
+      return c.json({
+        message: 'Student assignment not found',
+      }, 404)
+    }
+
+    studentAssignment = await prisma.student_assignments.update({
       where: {
         id: Number.parseInt(id),
       },
@@ -63,19 +84,36 @@ StudentAssignmentRouter
       },
     })
 
-    return c.json({ student_assignment: studentAssignment })
+    return c.json({
+      message: 'Updated',
+      student_assignment: studentAssignment,
+    })
   })
 
   // DELETE /api/student-assignments/:id
   .delete('/:id', async (c) => {
     const id = c.req.param('id')
-    const studentAssignment = await prisma.student_assignments.delete({
+    const studentAssignment = await prisma.student_assignments.findFirst({
       where: {
         id: Number.parseInt(id),
       },
     })
 
-    return c.json({ student_assignment: studentAssignment })
+    if (!studentAssignment) {
+      return c.json({
+        message: 'Student assignment not found',
+      }, 404)
+    }
+    await prisma.student_assignments.delete({
+      where: {
+        id: Number.parseInt(id),
+      },
+    })
+
+    return c.json({
+      message: 'Deleted',
+      student_assignment: studentAssignment,
+    })
   })
 
 export default StudentAssignmentRouter
